Split VersionUtilsTest so pure JVM tests leave Robolectric
